Refactor & Implement transformers NLP
