+++ app/main.py
@@ -1,5 +1,8 @@
-from nltk.sentiment.vader import SentimentIntensityAnalyzer
-from gnews.fetcher import Gnews
+from newsfetchers.gnews.gnews import Gnews
+from nlpanalyzers.transformers.analyzer import TransformerAnalyzer
+
+# Instantiate sentiment analyzer
+analyzer = TransformerAnalyzer()
 
 # Set variables
 language = "en"
@@ -7,49 +10,6 @@ country = "us"
 numMaxNews = 10
 search = "Microsoft"
 
-# Fetch Google News
-gnews = Gnews()
-articles = gnews.fetchNews(language, country, numMaxNews, search)
-
-# Analyze news
-print("Analyzing Google news...")
-for articleCounter in range(len(articles)):
-
-    # title
-    title = articles[articleCounter]['title']
-    print(f"title: {title}")
-
-    # description
-    description = articles[articleCounter]['description']
-    print(f"description: {description}")
-
-    # url
-    url = articles[articleCounter]['url']
-    print(f"url: {url}")
-
-    # publishedAt
-    publishedAt = articles[articleCounter]['publishedAt']
-    print(f"publishedAt: {publishedAt}")
-
-    # sourceName
-    sourceName = articles[articleCounter]['source']['name']
-    print(f"source.name: {sourceName}")
-
-    # sourceUrl
-    sourceUrl = articles[articleCounter]['source']['url']
-    print(f"source.url: {sourceUrl}")
-
-    # Analyze sentiment
-    sentimentAnalyzer = SentimentIntensityAnalyzer()
-    sentimentScores = sentimentAnalyzer.polarity_scores(title)
-
-    # Print sentiment scores
-    compound = sentimentScores["compound"]
-    neg = sentimentScores["neg"]
-    neu = sentimentScores["neu"]
-    pos = sentimentScores["pos"]
-
-    print(f"compound: {compound} | neu: {neu} | pos: {pos} | neg: {neg}")
-    print()
-
-print("Google news are analyzed successfully.")
+# Analyze Google News
+gnews = Gnews(analyzer)
+gnews.fetchAndAnalyzeArticles(language, country, numMaxNews, search)
