+++ app/nlpanalyzers/analyzer.py
@@ -1,0 +1,11 @@
+from abc import ABC, abstractmethod
+
+
+class NlpAnalyzer(ABC):
+
+    @abstractmethod
+    def analyzeStatement(
+        self,
+        statement,
+    ):
+        pass
