+++ app/requirements.txt
@@ -1,1 +1,2 @@
-nltk==3.8.1
+torch==2.0.1
+transformers==4.33.3
